Reduce logging to the repo's actual runtime contract
The previous log package carried context decoration, valuers, filters, and full-logger wrappers that the codebase no longer needed, while also making tests stateful and harder to reason about. This keeps only the minimal global logger surface and std logger implementation used by the repo.

Constraint: Preserve existing package-level Debug/Info/Warn/Error/Fatal call sites across transport, server, channel, and engine
Rejected: Keep context/valuer/filter layers and patch tests around them | extra abstraction had no in-repo consumers and created avoidable global-state coupling
Rejected: Replace package-level helpers with injected loggers everywhere | broader refactor than this cleanup requires
Directive: Keep the log package stateless apart from the single replaceable global logger unless a real in-repo requirement emerges
Confidence: high
Scope-risk: narrow
Tested: env GOCACHE=/tmp/go-build go test ./log -count=1
Tested: env GOCACHE=/tmp/go-build go test ./transport/tcp ./server ./internal/channel ./internal/engine . -count=1
Not-tested: go test ./... still has unrelated pre-existing failures outside log
