+++ log/log_test.go
@@ -2,62 +2,63 @@ package log
 
 import (
 	"bytes"
-	"context"
 	"fmt"
 	"strings"
 	"testing"
 )
 
 type mockLogger struct {
-	t *testing.T
+	entries []string
 }
 
 func (l *mockLogger) Log(level Level, kvs ...interface{}) {
-	builder := strings.Builder{}
-	builder.WriteString(level.String() + " ")
+	var builder strings.Builder
+	builder.WriteString(level.String())
 	for i := 0; i < len(kvs); i += 2 {
-		builder.WriteString(fmt.Sprintf("%v=%v ", kvs[i], kvs[i+1]))
+		builder.WriteString(fmt.Sprintf(" %v=%v", kvs[i], kvs[i+1]))
 	}
-	l.t.Log(builder.String())
+	l.entries = append(l.entries, builder.String())
 }
 
-func TestWith(t *testing.T) {
-	logger := Logger(&mockLogger{t: t})
-	logger, err := With(logger, "ts", DefaultTimestamp, "caller", DefaultCaller)
-	if err != nil {
-		t.Fatalf("expect: %v, got: %v", nil, err)
-	}
-	logger.Log(LevelDebug, "msg", "test1")
-	logger.Log(LevelInfo, "msg", "test2")
-	logger.Log(LevelError, "msg", "test3")
+func resetLogger(t *testing.T) {
+	t.Helper()
+	t.Cleanup(func() {
+		SetLogger(NewStdLogger(&bytes.Buffer{}))
+	})
+}
 
-	logger, err = With(logger, "singular")
-	if err == nil {
-		t.Fatalf("expect: %v, got: %v", ErrKvsNotInPaired, nil)
-		return
+func TestSetLoggerAndGetLogger(t *testing.T) {
+	resetLogger(t)
+
+	logger := &mockLogger{}
+	SetLogger(logger)
+
+	if got := GetLogger(); got != logger {
+		t.Fatal("GetLogger did not return the logger set by SetLogger")
+	}
+
+	Info("hello")
+	Debugw("key", "value")
+
+	expected := []string{
+		"INFO msg=hello",
+		"DEBUG key=value",
+	}
+	if len(logger.entries) != len(expected) {
+		t.Fatalf("got %v, want %v", logger.entries, expected)
+	}
+	for i := range expected {
+		if logger.entries[i] != expected[i] {
+			t.Fatalf("entry[%d] = %q, want %q", i, logger.entries[i], expected[i])
+		}
 	}
 }
 
-func TestWithContext(t *testing.T) {
-	logger := Logger(&mockLogger{t: t})
-	cnt := 0
-	type ctxCntKey struct{}
-	ctx := context.WithValue(context.Background(), ctxCntKey{}, &cnt)
-	logger, _ = WithContext(ctx, logger)
-	logger, _ = With(logger, "count", Valuer(func(ctx context.Context) interface{} {
-		cnt := ctx.Value(ctxCntKey{}).(*int)
-		*cnt++
-		return *cnt
-	}))
-	logger.Log(LevelDebug, "msg", "test1")
-	logger.Log(LevelInfo, "msg", "test2")
-	logger.Log(LevelError, "msg", "test3")
-}
+func TestPackageLoggerOutput(t *testing.T) {
+	resetLogger(t)
 
-func TestFullLogger(t *testing.T) {
 	buff := &bytes.Buffer{}
 	SetLogger(NewStdLogger(buff))
-	logger := NewContextLogger(context.Background())
 
 	cases := []struct {
 		level   Level
@@ -94,41 +95,57 @@ func TestFullLogger(t *testing.T) {
 		msg := fmt.Sprintf(c.content[0].(string), c.content[1:]...)
 		switch c.level {
 		case LevelDebug:
-			logger.Debug(msg)
+			Debug(msg)
 			expected = append(expected, fmt.Sprintf("%s msg=%s", LevelDebug.String(), msg))
-			logger.Debugf(c.content[0].(string), c.content[1:]...)
+			Debugf(c.content[0].(string), c.content[1:]...)
 			expected = append(expected, fmt.Sprintf("%s msg=%s", LevelDebug.String(), msg))
-			logger.Debugw("log", msg)
+			Debugw("log", msg)
 			expected = append(expected, fmt.Sprintf("%s log=%s", LevelDebug.String(), msg))
 		case LevelInfo:
-			logger.Info(msg)
+			Info(msg)
 			expected = append(expected, fmt.Sprintf("%s msg=%s", LevelInfo.String(), msg))
-			logger.Infof(c.content[0].(string), c.content[1:]...)
+			Infof(c.content[0].(string), c.content[1:]...)
 			expected = append(expected, fmt.Sprintf("%s msg=%s", LevelInfo.String(), msg))
-			logger.Infow("log", msg)
+			Infow("log", msg)
 			expected = append(expected, fmt.Sprintf("%s log=%s", LevelInfo.String(), msg))
 		case LevelWarn:
-			logger.Warn(msg)
+			Warn(msg)
 			expected = append(expected, fmt.Sprintf("%s msg=%s", LevelWarn.String(), msg))
-			logger.Warnf(c.content[0].(string), c.content[1:]...)
+			Warnf(c.content[0].(string), c.content[1:]...)
 			expected = append(expected, fmt.Sprintf("%s msg=%s", LevelWarn.String(), msg))
-			logger.Warnw("log", msg)
+			Warnw("log", msg)
 			expected = append(expected, fmt.Sprintf("%s log=%s", LevelWarn.String(), msg))
 		case LevelError:
-			logger.Error(msg)
+			Error(msg)
 			expected = append(expected, fmt.Sprintf("%s msg=%s", LevelError.String(), msg))
-			logger.Errorf(c.content[0].(string), c.content[1:]...)
+			Errorf(c.content[0].(string), c.content[1:]...)
 			expected = append(expected, fmt.Sprintf("%s msg=%s", LevelError.String(), msg))
-			logger.Errorw("log", msg)
+			Errorw("log", msg)
 			expected = append(expected, fmt.Sprintf("%s log=%s", LevelError.String(), msg))
 		}
 	}
 
+	Log(LevelInfo, DefaultMsgKey, "test log")
+	expected = append(expected, fmt.Sprintf("%s msg=%s", LevelInfo.String(), "test log"))
+
 	result := buff.String()
-	t.Log(result)
 	expected = append(expected, "")
 	if result != strings.Join(expected, "\n") {
 		t.Errorf("want: \n%s, got: \n%s", strings.Join(expected, "\n"), result)
 	}
+}
 
+func TestStdLoggerHandlesOddAndEmptyKeyvals(t *testing.T) {
+	logger := NewStdLogger(&bytes.Buffer{})
+	std := logger.(*stdLogger)
+
+	std.Log(LevelDebug)
+	if got := std.log.Writer().(*bytes.Buffer).String(); got != "" {
+		t.Fatalf("empty log call wrote %q, want empty", got)
+	}
+
+	std.Log(LevelDebug, "singular")
+	if got := std.log.Writer().(*bytes.Buffer).String(); got != "DEBUG singular=KEYVALS UNPAIRED\n" {
+		t.Fatalf("got %q", got)
+	}
 }
