+++ log/std.go
@@ -4,31 +4,24 @@ import (
 	"bytes"
 	"fmt"
 	"io"
-	"log"
-	"sync"
+	stdlog "log"
 )
 
-// COPY FROM https://github.com/go-kratos/kratos/blob/main/log/std.go
 var _ Logger = (*stdLogger)(nil)
 
 type stdLogger struct {
-	log  *log.Logger
-	pool *sync.Pool
+	log *stdlog.Logger
 }
 
-// NewStdLogger new a logger with writer.
 func NewStdLogger(w io.Writer) Logger {
+	if w == nil {
+		w = io.Discard
+	}
 	return &stdLogger{
-		log: log.New(w, "", 0),
-		pool: &sync.Pool{
-			New: func() interface{} {
-				return new(bytes.Buffer)
-			},
-		},
+		log: stdlog.New(w, "", 0),
 	}
 }
 
-// Log print the kv pairs log.
 func (l *stdLogger) Log(level Level, keyvals ...interface{}) {
 	if len(keyvals) == 0 {
 		return
@@ -36,12 +29,10 @@ func (l *stdLogger) Log(level Level, keyvals ...interface{}) {
 	if (len(keyvals) & 1) == 1 {
 		keyvals = append(keyvals, "KEYVALS UNPAIRED")
 	}
-	buf := l.pool.Get().(*bytes.Buffer)
+	var buf bytes.Buffer
 	buf.WriteString(level.String())
 	for i := 0; i < len(keyvals); i += 2 {
-		_, _ = fmt.Fprintf(buf, " %s=%v", keyvals[i], keyvals[i+1])
+		_, _ = fmt.Fprintf(&buf, " %v=%v", keyvals[i], keyvals[i+1])
 	}
 	_ = l.log.Output(4, buf.String())
-	buf.Reset()
-	l.pool.Put(buf)
 }
