--- log/value.go
@@ -1,66 +1,0 @@
-package log
-
-import (
-	"context"
-	"runtime"
-	"strconv"
-	"strings"
-	"time"
-)
-
-// COPY FROM https://github.com/go-kratos/kratos/blob/main/log/value.go
-var (
-	// DefaultCaller is a Valuer that returns the file and line.
-	DefaultCaller = Caller(4)
-
-	// DefaultTimestamp is a Valuer that returns the current wallclock time.
-	DefaultTimestamp = Timestamp(time.RFC3339)
-)
-
-// Valuer is returns a log value.
-type Valuer func(ctx context.Context) interface{}
-
-// Value return the function value.
-func Value(ctx context.Context, v interface{}) interface{} {
-	if v, ok := v.(Valuer); ok {
-		return v(ctx)
-	}
-	return v
-}
-
-// Caller returns a Valuer that returns a pkg/file:line description of the caller.
-func Caller(depth int) Valuer {
-	return func(context.Context) interface{} {
-		_, file, line, _ := runtime.Caller(depth)
-		idx := strings.LastIndexByte(file, '/')
-		if idx == -1 {
-			return file[idx+1:] + ":" + strconv.Itoa(line)
-		}
-		idx = strings.LastIndexByte(file[:idx], '/')
-		return file[idx+1:] + ":" + strconv.Itoa(line)
-	}
-}
-
-// Timestamp returns a timestamp Valuer with a custom time format.
-func Timestamp(layout string) Valuer {
-	return func(context.Context) interface{} {
-		return time.Now().Format(layout)
-	}
-}
-
-func calculateValues(ctx context.Context, keyvals []interface{}) {
-	for i := 1; i < len(keyvals); i += 2 {
-		if v, ok := keyvals[i].(Valuer); ok {
-			keyvals[i] = v(ctx)
-		}
-	}
-}
-
-func containsValuer(keyvals []interface{}) bool {
-	for i := 1; i < len(keyvals); i += 2 {
-		if _, ok := keyvals[i].(Valuer); ok {
-			return true
-		}
-	}
-	return false
-}
